// frontend/sdk/models/DailyTodayDto.ts
export interface DailyTodayDto {
    /**
     * The ID of the game session
     * @type {string}
     * @memberof DailyTodayDto
     */
    sessionId: string;
    /**
     * The current round of the game
     * @type {number}
     * @memberof DailyTodayDto
     */
    currentRound: number;
    /**
     * The duration of the current snippet
     * @type {number}
     * @memberof DailyTodayDto
     */
    snippetDuration: number;
    /**
     * The status of the game
     * @type {string}
     * @memberof DailyTodayDto
     */
    status: DailyTodayDtoStatusEnum;
    /**
     * The guesses of the game
     * @type {Array<GuessHistoryDto>}
     * @memberof DailyTodayDto
     */
    guesses: Array<GuessHistoryDto>;
    /**
     * The preview URL of the current track
     * @type {string}
     * @memberof DailyTodayDto
     */
    previewUrl?: string;
    /**
     * The answer to the current track
     * @type {TrackOptionDto}
     * @memberof DailyTodayDto
     */
    answer?: TrackOptionDto;
    /**
     * Track options to choose from (from same playlist(s) as the daily)
     * @type {Array<TrackOptionDto>}
     * @memberof DailyTodayDto
     */
    trackOptions: Array<TrackOptionDto>;
    /**
     * Playlist name for this daily
     * @type {string}
     * @memberof DailyTodayDto
     */
    playlistName?: string;
}


/**
 * @export
 */
export const DailyTodayDtoStatusEnum = {
    Playing: 'PLAYING',
    Won: 'WON',
    Lost: 'LOST'
} as const;
export type DailyTodayDtoStatusEnum = typeof DailyTodayDtoStatusEnum[keyof typeof DailyTodayDtoStatusEnum];